requirements.txt
beautifulsoup4~=4.12.2
folium
matplotlib
numpy
pandas>=2.0.3
Pillow>=9.4.0
Pint
plotly
psycopg[binary]
requests~=2.31.0
sshtunnel~=0.4.0
sympy~=1.12
FreeSimpleGUI~=5.0.0
html5lib~=1.1
lxml
soupsieve~=2.4.1
jupyterlab
defusedxml~=0.7.1
cffi>=1.15.1
future~=0.18.3
ipython~=8.12.0
Babel~=2.12.1
dask
pytz~=2023.3
mpmath~=1.3.0
latex~=0.7.0
decorator~=5.1.1
Jinja2~=3.1.2
branca~=0.6.0
MarkupSafe~=2.1.2
openpyxl~=3.1.2
python-dateutil~=2.8.2
scipy>=1.9.1
certifi~=2022.12.7
cryptography~=40.0.2
six~=1.16.0
paramiko~=3.1.0
bcrypt~=4.0.1
gssapi>=1.8.2
idna~=3.4
urllib3~=1.26.15
simplejson~=3.19.1
contourpy>=1.0.7
fonttools~=4.39.3
reportlab~=4.0.0
pycparser~=2.21
pyparsing~=3.0.9
deprecation~=2.1.0
cycler~=0.11.0
packaging~=23.1
ipykernel~=6.22.0
tornado~=6.3
docutils~=0.20
kiwisolver~=1.4.4
zipp~=3.15.0
build~=0.10.0
wheel~=0.38.4
tifftools~=1.4.0
natsort~=8.4.0
pyntcloud~=0.3.1
kivymd~=1.1.1
PyQt5~=5.15.9
kivy>=2.2.1
termcolor~=2.3.0
pytest-cov~=4.1.0
pytest~=7.4.1
setuptools~=68.2.2
nbformat~=5.9.2
traitlets~=5.13.0
nbconvert~=7.16.3
pymupdf~=1.24.2
flask~=3.0.0
json2html~=1.3.0
werkzeug~=3.0.1
sqlalchemy~=2.0.30
civilpy~=0.1.1
tqdm~=4.66.1
streamlit~=1.37.0
dill~=0.3.8
geopandas~=0.13.2
markdownify~=0.13.1